# docs/android-handoff/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SwaggerExport",
    "version": "1.0"
  },
  "paths": {
...
    "/api/DocChiSo/P_00_KET_NOI_DB_CHECK": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_00_KET_NOI_DB_CHECKRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_00_KET_NOI_DB_CHECKRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_00_KET_NOI_DB_CHECKRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_01_DANG_NHAP": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_01_DANG_NHAPRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_01_DANG_NHAPRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_01_DANG_NHAPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_011_DANG_NHAP_DOI_MAT_KHAU": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_011_DANG_NHAP_DOI_MAT_KHAURequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_011_DANG_NHAP_DOI_MAT_KHAURequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_011_DANG_NHAP_DOI_MAT_KHAURequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_012_LAY_GT_CANH_BAO": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_012_LAY_GT_CANH_BAORequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_012_LAY_GT_CANH_BAORequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_012_LAY_GT_CANH_BAORequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_013_LAY_PHIEN_BAN_APP": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_013_LAY_PHIEN_BAN_APPRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_013_LAY_PHIEN_BAN_APPRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_013_LAY_PHIEN_BAN_APPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_01_DANG_NHAP_LUU_TOKEN": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_01_DANG_NHAP_LUU_TOKENRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_01_DANG_NHAP_LUU_TOKENRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_01_DANG_NHAP_LUU_TOKENRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_021_LAY_DS_SO_DOC": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_021_LAY_DS_SO_DOCRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_021_LAY_DS_SO_DOCRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_021_LAY_DS_SO_DOCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_0211_LAY_DS_SO_DOC_BS": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_0211_LAY_DS_SO_DOC_BSRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_0211_LAY_DS_SO_DOC_BSRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_0211_LAY_DS_SO_DOC_BSRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_0212_LAY_DS_SO_DOC_TRA_CUU": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_0212_LAY_DS_SO_DOC_TRA_CUURequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_0212_LAY_DS_SO_DOC_TRA_CUURequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_0212_LAY_DS_SO_DOC_TRA_CUURequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_022_NHAN_SO_DOC": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_022_NHAN_SO_DOCRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_022_NHAN_SO_DOCRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_022_NHAN_SO_DOCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_03_LAY_DS_KHACH_HANG": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANGRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANGRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANGRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_03_LAY_DS_KHACH_HANG_SUB": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANG_SUBRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANG_SUBRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANG_SUBRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_03_LAY_DS_KHACH_HANG_SUB_BS": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANG_SUB_BSRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANG_SUB_BSRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_03_LAY_DS_KHACH_HANG_SUB_BSRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_0313_LAY_DS_KHACH_HANG_TRA_CUU": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_0313_LAY_DS_KHACH_HANG_TRA_CUURequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_0313_LAY_DS_KHACH_HANG_TRA_CUURequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_0313_LAY_DS_KHACH_HANG_TRA_CUURequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_043_LO_TRINH_DI_DOC_MAP": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_043_LO_TRINH_DI_DOC_MAPRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_043_LO_TRINH_DI_DOC_MAPRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_043_LO_TRINH_DI_DOC_MAPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_044_LUU_TEN_FILE_ANH": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_044_LUU_TEN_FILE_ANHRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_044_LUU_TEN_FILE_ANHRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_044_LUU_TEN_FILE_ANHRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_032_LAY_SL_BINH_QUAN_3T": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_032_LAY_SL_BINH_QUAN_3TRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_032_LAY_SL_BINH_QUAN_3TRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_032_LAY_SL_BINH_QUAN_3TRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_041_NHAP_XUAT_CS_LE_ONLINE": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_041_NHAP_XUAT_CS_LE_ONLINERequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_041_NHAP_XUAT_CS_LE_ONLINERequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_041_NHAP_XUAT_CS_LE_ONLINERequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_045_CANH_BAO_SAN_LUONG_LON": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_045_CANH_BAO_SAN_LUONG_LONRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_045_CANH_BAO_SAN_LUONG_LONRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_045_CANH_BAO_SAN_LUONG_LONRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_045_LUU_XEP_SO_DOC": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_045_LUU_XEP_SO_DOCRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_045_LUU_XEP_SO_DOCRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_045_LUU_XEP_SO_DOCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_041_NHAP_XUAT_CS_LE_ONLINE_SUB": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_041_NHAP_XUAT_CS_LE_ONLINE_SUBRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_041_NHAP_XUAT_CS_LE_ONLINE_SUBRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_041_NHAP_XUAT_CS_LE_ONLINE_SUBRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DocChiSo/P_051_KIEM_TRA_BAN_GIAO_SD": {
      "post": {
        "tags": [
          "DocChiSo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P_051_KIEM_TRA_BAN_GIAO_SDRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/P_051_KIEM_TRA_BAN_GIAO_SDRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/P_051_KIEM_TRA_BAN_GIAO_SDRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "P_00_KET_NOI_DB_CHECKRequest": {
        "type": "object",
        "properties": {
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "P_011_DANG_NHAP_DOI_MAT_KHAURequest": {
        "type": "object",
        "properties": {
          "MA_BIEN_DOC": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_OLD": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_NEW1": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_NEW2": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "P_012_LAY_GT_CANH_BAORequest": {
        "type": "object",
        "properties": {
          "MA_BIEN_DOC": {
            "type": "string",
            "nullable": true
          },
          "MA_XI_NGHIEP": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "P_01_DANG_NHAPRequest": {
        "type": "object",
        "properties": {
          "USER": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}